Fix signed/unsigned comparison warnings.

+++ lib/libc/rpc/getnetconfig.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: getnetconfig.c,v 1.8 2002/11/08 00:13:07 fvdl Exp $	*/
+/*	$NetBSD: getnetconfig.c,v 1.9 2002/11/11 17:13:59 thorpej Exp $	*/
 
 /*
  * Sun RPC is a product of Sun Microsystems, Inc. and is provided for
@@ -45,6 +45,7 @@ static        char sccsid[] = "@(#)getnetconfig.c	1.12 91/12/19 SMI";
 #include <assert.h>
 #include <errno.h>
 #include <netconfig.h>
+#include <stddef.h>
 #include <stdlib.h>
 #include <string.h>
 #include <rpc/rpc.h>
@@ -448,7 +449,7 @@ getnetconfigent(netid)
 		return (NULL);
 	}
 	do {
-		int len;
+		ptrdiff_t len;
 		char *tmpp;	/* tmp string pointer */
 
 		do {
@@ -463,7 +464,7 @@ getnetconfigent(netid)
 			nc_error = NC_BADFILE;
 			break;
 		}
-		if (strlen(netid) == (len = tmpp - stringp) &&	/* a match */
+		if (strlen(netid) == (size_t) (len = tmpp - stringp) &&	/* a match */
 		    strncmp(stringp, netid, (size_t)len) == 0) {
 			if ((ncp = (struct netconfig *)
 			    malloc(sizeof (struct netconfig))) == NULL)
@@ -645,7 +646,7 @@ static struct netconfig *
 {
 	struct netconfig	*p;
 	char	*tmp;
-	int	i;
+	u_int	i;
 
 	_DIAGASSERT(ncp != NULL);
